+++ lib/version-headers.ts
@@ -1,0 +1,5 @@
+import { getAppVersion } from "./version";
+
+export function forgeguardVersionHeaders(): HeadersInit {
+  return { "X-ForgeGuard-Version": getAppVersion() };
+}